Mid of November 2025 update.
This small update aims to mostly bring some features that were holding still, which improving various aspects of our functions.

The changes are listed here:

- Added LZ support for SDAT unpacking,
- Optimized base64 decoding for unbar and bar (no duplication),
- Bundle the XAMPP PHP installation with ApacheNet,
- Implements Mitm server selector for Blaze based solutions in MultiSocks,
- Merged Tcp and Udp utils into a single class for future experiments,
- Implements back the WarhawkRevived Medius timeout logic (to fix clock overflow that is occuring on PS3 clients),
- Fixes a crashing issue while rebooting Horizon servers,
- Removed useless helper funcs that bloated code,
- Fixed a AFSMapper crashing issue,
- Various code styles improvements and unification.

+++ Servers/ApacheNet/BuildIn/Extensions/PHP.cs
@@ -36,13 +36,13 @@ namespace ApacheNet.BuildIn.Extensions
             string? documentRootPath = Path.GetDirectoryName(FilePath);
             string? scriptFilePath = Path.GetFullPath(FilePath);
             string? scriptFileName = Path.GetFileName(FilePath);
+            string phpFullPath = $"{PHPPath}/{PHPVer}/";
 
             PostData = ctx.Request.DataAsBytes;
 
-            TheProcess.StartInfo.FileName = $"{PHPPath}/{PHPVer}/php-cgi";
+            TheProcess.StartInfo.FileName = $"{phpFullPath}php-cgi";
 
-            TheProcess.StartInfo.Arguments = $"-q -c \"{$"{PHPPath}/{PHPVer}/php.ini"}\" -d \"error_reporting=E_ALL\" -d \"display_errors={ApacheNetServerConfiguration.PHPDebugErrors}\" -d \"expose_php=Off\" -d \"include_path='{documentRootPath}'\" " +
-                         $"-d \"extension_dir='{$"{PHPPath}/{PHPVer}/ext/"}'\" \"{FilePath}\"";
+            TheProcess.StartInfo.Arguments = $"-q -c \"{$"{phpFullPath}php.ini"}\" -d \"error_reporting=E_ALL\" -d \"display_errors={ApacheNetServerConfiguration.PHPDebugErrors}\" -d \"expose_php=Off\" -d \"include_path='{documentRootPath}'\" \"{FilePath}\"";
 
             TheProcess.StartInfo.CreateNoWindow = true;
             TheProcess.StartInfo.UseShellExecute = false;
@@ -60,8 +60,8 @@ namespace ApacheNet.BuildIn.Extensions
             TheProcess.StartInfo.EnvironmentVariables["SYSTEMROOT"] = Environment.GetEnvironmentVariable("SYSTEMROOT");
             TheProcess.StartInfo.EnvironmentVariables["WINDIR"] = Environment.GetEnvironmentVariable("WINDIR");
             TheProcess.StartInfo.EnvironmentVariables["COMSPEC"] = Environment.GetEnvironmentVariable("COMSPEC");
-            TheProcess.StartInfo.EnvironmentVariables["TMPDIR"] = Environment.GetEnvironmentVariable("TMPDIR");
-            TheProcess.StartInfo.EnvironmentVariables["TEMP"] = Environment.GetEnvironmentVariable("TEMP");
+            TheProcess.StartInfo.EnvironmentVariables["TMPDIR"] = Environment.GetEnvironmentVariable("TMPDIR") ?? Path.GetTempPath();
+            TheProcess.StartInfo.EnvironmentVariables["TEMP"] = Environment.GetEnvironmentVariable("TEMP") ?? Path.GetTempPath();
             TheProcess.StartInfo.EnvironmentVariables["PATH"] = Environment.GetEnvironmentVariable("PATH");
             TheProcess.StartInfo.EnvironmentVariables.Add("GATEWAY_INTERFACE", "CGI/1.1");
             TheProcess.StartInfo.EnvironmentVariables.Add("SERVER_PROTOCOL", $"HTTP/{ApacheNetServerConfiguration.HttpVersion}");
@@ -145,8 +145,19 @@ namespace ApacheNet.BuildIn.Extensions
             catch (Exception ex)
             {
                 CustomLogger.LoggerAccessor.LogWarn($"[PHP] - Killing bad process. (Exception:{ex})");
-                TheProcess.Kill();
-                TheProcess.WaitForExit();
+
+                try
+                {
+                    if (TheProcess != null && !TheProcess.HasExited)
+                    {
+                        TheProcess.Kill();
+                        TheProcess.WaitForExit();
+                    }
+                }
+                catch (Exception innerEx)
+                {
+                    CustomLogger.LoggerAccessor.LogError($"[PHP] - Failed to kill process. (Exception:{innerEx})");
+                }
             }
         }
